levo-schema-service/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
typer
pydantic
pyyaml
openapi-spec-validator
pytest
httpx